Use Chunks in RateLimiter metrics
